优化页面 选择第一个笔记

+++ src/components/TrashDetail.vue
@@ -48,10 +48,13 @@ export default {
   created() {
     this.checkLogin({path: '/login'})
     this.getNotebooks()
-    this.getTrashNotes()
-      .then(() => {
-        this.setCurTrashNote({curTrashNoteId: parseInt(this.$route.query.noteId)})
+    this.getTrashNotes().then(() => {
+      this.setCurTrashNote({curTrashNoteId: this.$route.query.noteId})
+      this.$router.replace({
+        path: '/trash',
+        query: {noteId: this.curTrashNote.id}
       })
+    })
   },
   computed: {
     ...mapGetters([
@@ -75,12 +78,29 @@ export default {
       'getNotebooks'
     ]),
     onDelete() {
-      this.deleteTrashNote({noteId: this.curTrashNote.id})
-      console.log('彻底删除')
+      this.$confirm('删除后将无法恢复', '确定删除？', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        return this.deleteTrashNote({noteId: this.curTrashNote.id})
+      }).then(() => {
+        this.setCurTrashNote()
+        this.$router.replace({
+          path: '/trash',
+          query: {noteId: this.curTrashNote.id}
+        })
+      })
     },
     onRevert() {
       this.revertTrashNote({noteId: this.curTrashNote.id})
-      console.log('恢复')
+        .then(() => {
+          this.setCurTrashNote()
+          this.$router.replace({
+            path: '/trash',
+            query: {noteId: this.curTrashNote.id}
+          })
+        })
     }
   },
   beforeRouteUpdate(to, from, next) {
